feat: forming an encapsulation pattern
