+++ src/notebooks/pipes/Visualization/types/Heatmap/icon.tsx
@@ -1,0 +1,241 @@
+import React from 'react'
+
+const icon = (
+  <div className="vis-graphic" data-testid="vis-graphic--heatmap">
+    <svg
+      width="100%"
+      height="100%"
+      version="1.1"
+      id="Histogram"
+      x="0px"
+      y="0px"
+      viewBox="0 0 150 150"
+      preserveAspectRatio="none meet"
+    >
+      <g id="r">
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-b"
+          points="127.5,127.5 127.5,112.5 112.5,112.5 112.5,127.5 97.5,127.5 97.5,142.5 112.5,142.5 127.5,142.5
+		142.5,142.5 142.5,127.5 	"
+        />
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-b"
+          points="67.5,127.5 52.5,127.5 52.5,112.5 37.5,112.5 37.5,97.5 22.5,97.5 22.5,112.5 22.5,127.5 7.5,127.5
+		7.5,142.5 22.5,142.5 37.5,142.5 52.5,142.5 67.5,142.5 82.5,142.5 82.5,127.5 82.5,112.5 67.5,112.5 	"
+        />
+      </g>
+      <g id="rl">
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-b"
+            d="M126,114v13.5v1.5h1.5H141v12h-13.5h-15H99v-12h13.5h1.5v-1.5V114H126 M127.5,112.5h-15v15h-15v15h15h15h15
+			v-15h-15V112.5L127.5,112.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-b"
+            d="M36,99v13.5v1.5h1.5H51v13.5v1.5h1.5h15H69v-1.5V114h12v13.5V141H67.5h-15h-15h-15H9v-12h13.5H24v-1.5v-15V99
+			H36 M37.5,97.5h-15v15v15h-15v15h15h15h15h15h15v-15v-15h-15v15h-15v-15h-15V97.5L37.5,97.5z"
+          />
+        </g>
+      </g>
+      <g id="g">
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-c"
+          points="97.5,97.5 97.5,82.5 82.5,82.5 82.5,97.5 67.5,97.5 52.5,97.5 37.5,97.5 37.5,112.5 52.5,112.5
+		52.5,127.5 67.5,127.5 67.5,112.5 82.5,112.5 82.5,127.5 82.5,142.5 97.5,142.5 97.5,127.5 112.5,127.5 112.5,112.5 97.5,112.5
+		"
+        />
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-c"
+          points="127.5,82.5 127.5,97.5 112.5,97.5 112.5,112.5 127.5,112.5 127.5,127.5 142.5,127.5 142.5,112.5
+		142.5,97.5 142.5,82.5 	"
+        />
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-c"
+          points="37.5,67.5 22.5,67.5 22.5,82.5 7.5,82.5 7.5,97.5 7.5,112.5 7.5,127.5 22.5,127.5 22.5,112.5
+		22.5,97.5 37.5,97.5 37.5,82.5 	"
+        />
+      </g>
+      <g id="gl">
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-c"
+            d="M96,84v13.5v15v1.5h1.5H111v12H97.5H96v1.5V141H84v-13.5v-15V111h-1.5h-15H66v1.5V126H54v-13.5V111h-1.5H39
+			V99h13.5h15h15H84v-1.5V84H96 M97.5,82.5h-15v15h-15h-15h-15v15h15v15h15v-15h15v15v15h15v-15h15v-15h-15v-15V82.5L97.5,82.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-c"
+            d="M141,84v13.5v15V126h-12v-13.5V111h-1.5H114V99h13.5h1.5v-1.5V84H141 M142.5,82.5h-15v15h-15v15h15v15h15v-15
+			v-15V82.5L142.5,82.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-c"
+            d="M36,69v13.5V96H22.5H21v1.5v15V126H9v-13.5v-15V84h13.5H24v-1.5V69H36 M37.5,67.5h-15v15h-15v15v15v15h15v-15
+			v-15h15v-15V67.5L37.5,67.5z"
+          />
+        </g>
+      </g>
+      <g id="b">
+        <rect
+          x="67.5"
+          y="82.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-a"
+          points="52.5,67.5 37.5,67.5 37.5,82.5 37.5,97.5 52.5,97.5 52.5,82.5 67.5,82.5 67.5,67.5 	"
+        />
+        <rect
+          x="127.5"
+          y="67.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-a"
+          points="22.5,52.5 7.5,52.5 7.5,67.5 22.5,67.5 37.5,67.5 37.5,52.5 	"
+        />
+        <rect
+          x="67.5"
+          y="52.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <polygon
+          className="vis-graphic--fill vis-graphic--fill-a"
+          points="112.5,67.5 97.5,67.5 82.5,67.5 82.5,82.5 97.5,82.5 97.5,97.5 97.5,112.5 112.5,112.5 112.5,97.5
+		127.5,97.5 127.5,82.5 112.5,82.5 	"
+        />
+        <rect
+          x="82.5"
+          y="37.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <rect
+          x="112.5"
+          y="52.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <rect
+          x="37.5"
+          y="37.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <rect
+          x="127.5"
+          y="37.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <rect
+          x="22.5"
+          y="22.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+        <rect
+          x="67.5"
+          y="7.5"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="15"
+          height="15"
+        />
+      </g>
+      <g id="bl">
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M81,84v12H69V84H81 M82.5,82.5h-15v15h15V82.5L82.5,82.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M66,69v12H52.5H51v1.5V96H39V82.5V69h13.5H66 M67.5,67.5h-15h-15v15v15h15v-15h15V67.5L67.5,67.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M141,69v12h-12V69H141 M142.5,67.5h-15v15h15V67.5L142.5,67.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M36,54v12H22.5H9V54h13.5H36 M37.5,52.5h-15h-15v15h15h15V52.5L37.5,52.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M81,54v12H69V54H81 M82.5,52.5h-15v15h15V52.5L82.5,52.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M111,69v13.5V84h1.5H126v12h-13.5H111v1.5V111H99V97.5v-15V81h-1.5H84V69h13.5H111 M112.5,67.5h-15h-15v15h15
+			v15v15h15v-15h15v-15h-15V67.5L112.5,67.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M96,39v12H84V39H96 M97.5,37.5h-15v15h15V37.5L97.5,37.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M126,54v12h-12V54H126 M127.5,52.5h-15v15h15V52.5L127.5,52.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M51,39v12H39V39H51 M52.5,37.5h-15v15h15V37.5L52.5,37.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M141,39v12h-12V39H141 M142.5,37.5h-15v15h15V37.5L142.5,37.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M36,24v12H24V24H36 M37.5,22.5h-15v15h15V22.5L37.5,22.5z"
+          />
+        </g>
+        <g>
+          <path
+            className="vis-graphic--fill vis-graphic--fill-a"
+            d="M81,9v12H69V9H81 M82.5,7.5h-15v15h15V7.5L82.5,7.5z"
+          />
+        </g>
+      </g>
+    </svg>
+  </div>
+)
+
+export default icon
