+++ src/notebooks/pipes/Visualization/types/Histogram/icon.tsx
@@ -1,0 +1,233 @@
+import React from 'react'
+
+const icon = (
+  <div className="vis-graphic" data-testid="vis-graphic--histogram">
+    <svg
+      width="100%"
+      height="100%"
+      version="1.1"
+      id="Histogram"
+      x="0px"
+      y="0px"
+      viewBox="0 0 150 150"
+      preserveAspectRatio="none meet"
+    >
+      <g>
+        <rect
+          x="122"
+          y="51"
+          className="vis-graphic--fill vis-graphic--fill-c"
+          width="26"
+          height="24"
+        />
+        <rect
+          x="122"
+          y="113"
+          className="vis-graphic--fill vis-graphic--fill-b"
+          width="26"
+          height="13"
+        />
+        <rect
+          x="122"
+          y="79"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="26"
+          height="30"
+        />
+        <rect
+          x="92"
+          y="74"
+          className="vis-graphic--fill vis-graphic--fill-b"
+          width="26"
+          height="52"
+        />
+        <rect
+          x="92"
+          y="49"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="26"
+          height="21"
+        />
+        <rect
+          x="62"
+          y="94"
+          className="vis-graphic--fill vis-graphic--fill-b"
+          width="26"
+          height="32"
+        />
+        <rect
+          x="62"
+          y="59"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="26"
+          height="31"
+        />
+        <rect
+          x="62"
+          y="24"
+          className="vis-graphic--fill vis-graphic--fill-c"
+          width="26"
+          height="31"
+        />
+        <rect
+          x="32"
+          y="116"
+          className="vis-graphic--fill vis-graphic--fill-b"
+          width="26"
+          height="10"
+        />
+        <rect
+          x="32"
+          y="80"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="26"
+          height="32"
+        />
+        <rect
+          x="2"
+          y="110"
+          className="vis-graphic--fill vis-graphic--fill-b"
+          width="26"
+          height="16"
+        />
+        <rect
+          x="2"
+          y="90"
+          className="vis-graphic--fill vis-graphic--fill-a"
+          width="26"
+          height="16"
+        />
+        <rect
+          x="32"
+          y="64"
+          className="vis-graphic--fill vis-graphic--fill-c"
+          width="26"
+          height="12"
+        />
+        <rect
+          x="92"
+          y="35"
+          className="vis-graphic--fill vis-graphic--fill-c"
+          width="26"
+          height="10"
+        />
+        <rect
+          x="2"
+          y="81"
+          className="vis-graphic--fill vis-graphic--fill-c"
+          width="26"
+          height="5"
+        />
+      </g>
+      <g>
+        <rect
+          x="122"
+          y="51"
+          className="vis-graphic--line vis-graphic--line-c"
+          width="26"
+          height="24"
+        />
+        <rect
+          x="122"
+          y="113"
+          className="vis-graphic--line vis-graphic--line-b"
+          width="26"
+          height="13"
+        />
+        <rect
+          x="122"
+          y="79"
+          className="vis-graphic--line vis-graphic--line-a"
+          width="26"
+          height="30"
+        />
+        <rect
+          x="92"
+          y="74"
+          className="vis-graphic--line vis-graphic--line-b"
+          width="26"
+          height="52"
+        />
+        <rect
+          x="92"
+          y="49"
+          className="vis-graphic--line vis-graphic--line-a"
+          width="26"
+          height="21"
+        />
+        <rect
+          x="62"
+          y="94"
+          className="vis-graphic--line vis-graphic--line-b"
+          width="26"
+          height="32"
+        />
+        <rect
+          x="62"
+          y="59"
+          className="vis-graphic--line vis-graphic--line-a"
+          width="26"
+          height="31"
+        />
+        <rect
+          x="62"
+          y="24"
+          className="vis-graphic--line vis-graphic--line-c"
+          width="26"
+          height="31"
+        />
+        <rect
+          x="32"
+          y="116"
+          className="vis-graphic--line vis-graphic--line-b"
+          width="26"
+          height="10"
+        />
+        <rect
+          x="32"
+          y="80"
+          className="vis-graphic--line vis-graphic--line-a"
+          width="26"
+          height="32"
+        />
+        <rect
+          x="2"
+          y="110"
+          className="vis-graphic--line vis-graphic--line-b"
+          width="26"
+          height="16"
+        />
+        <rect
+          x="2"
+          y="90"
+          className="vis-graphic--line vis-graphic--line-a"
+          width="26"
+          height="16"
+        />
+        <rect
+          x="32"
+          y="64"
+          className="vis-graphic--line vis-graphic--line-c"
+          width="26"
+          height="12"
+        />
+        <rect
+          x="92"
+          y="35"
+          className="vis-graphic--line vis-graphic--line-c"
+          width="26"
+          height="10"
+        />
+        <rect
+          x="2"
+          y="81"
+          className="vis-graphic--line vis-graphic--line-c"
+          width="26"
+          height="5"
+        />
+      </g>
+    </svg>
+  </div>
+)
+
+export default icon
