+++ src/notebooks/pipes/Visualization/types/Map/index.tsx
@@ -1,60 +1,5 @@
-import React from 'react'
-
-const icon = (
-  <div className="vis-graphic" data-testid="vis-graphic--mosaic">
-    <svg
-      width="100%"
-      height="100%"
-      version="1.1"
-      id="Bar"
-      x="0px"
-      y="0px"
-      viewBox="0 0 150 150"
-      preserveAspectRatio="none meet"
-    >
-      <polygon
-        className="vis-graphic--fill vis-graphic--fill-a"
-        points="15,100.5 42.3,114.1 69.5,100.5 42.3,86.9 "
-      />
-      <polygon
-        className="vis-graphic--fill vis-graphic--fill-a"
-        points="80.5,100.5 107.7,114.1 135,100.5 107.7,86.9 "
-      />
-      <polygon
-        className="vis-graphic--fill vis-graphic--fill-a"
-        points="47.7,84.1 75,97.8 102.3,84.1 75,70.5 "
-      />
-      <polygon
-        className="vis-graphic--fill vis-graphic--fill-a"
-        points="47.7,116.9 75,130.5 102.3,116.9 75,103.2 "
-      />
-      <polygon
-        className="vis-graphic--line vis-graphic-line-a"
-        points="15,100.5 42.3,114.1 69.5,100.5 42.3,86.9 "
-      />
-      <polygon
-        className="vis-graphic--line vis-graphic-line-a"
-        points="80.5,100.5 107.7,114.1 135,100.5 107.7,86.9 "
-      />
-      <polygon
-        className="vis-graphic--line vis-graphic-line-a"
-        points="47.7,84.1 75,97.8 102.3,84.1 75,70.5 "
-      />
-      <polygon
-        className="vis-graphic--line vis-graphic-line-a"
-        points="47.7,116.9 75,130.5 102.3,116.9 75,103.2 "
-      />
-      <path
-        className="vis-graphic--fill vis-graphic--fill-b"
-        d="M92.7,26.8c-9.8-9.8-25.6-9.8-35.4,0s-9.8,25.6,0,35.4L75,79.9l17.7-17.7C102.4,52.4,102.4,36.6,92.7,26.8z M82.8,52.3c-4.3,4.3-11.3,4.3-15.6,0c-4.3-4.3-4.3-11.3,0-15.6c4.3-4.3,11.3-4.3,15.6,0C87.1,41,87.1,48,82.8,52.3z"
-      />
-      <path
-        className="vis-graphic--line vis-graphic--line-b"
-        d="M92.7,26.8c-9.8-9.8-25.6-9.8-35.4,0s-9.8,25.6,0,35.4L75,79.9l17.7-17.7C102.4,52.4,102.4,36.6,92.7,26.8z M82.8,52.3c-4.3,4.3-11.3,4.3-15.6,0c-4.3-4.3-4.3-11.3,0-15.6c4.3-4.3,11.3-4.3,15.6,0C87.1,41,87.1,48,82.8,52.3z"
-      />
-    </svg>
-  </div>
-)
+import icon from './icon'
+import properties from './properties'
 
 export default register => {
   register({
@@ -62,32 +7,6 @@ export default register => {
     name: 'Map',
     graphic: icon,
     disabled: true,
-    initial: {
-      type: 'xy',
-      position: 'overlaid',
-      legend: {},
-      note: '',
-      showNoteWhenEmpty: false,
-      axes: {
-        x: {
-          bounds: ['', ''],
-          label: '',
-          prefix: '',
-          suffix: '',
-          base: '10',
-          scale: 'linear',
-        },
-        y: {
-          bounds: ['', ''],
-          label: '',
-          prefix: '',
-          suffix: '',
-          base: '10',
-          scale: 'linear',
-        },
-      },
-      geom: 'line',
-      shape: 'chronograf-v2',
-    },
+    initial: properties,
   })
 }
