+++ src/notebooks/pipes/Visualization/types/Map/properties.ts
@@ -1,0 +1,27 @@
+export default {
+  type: 'xy',
+  position: 'overlaid',
+  legend: {},
+  note: '',
+  showNoteWhenEmpty: false,
+  axes: {
+    x: {
+      bounds: ['', ''],
+      label: '',
+      prefix: '',
+      suffix: '',
+      base: '10',
+      scale: 'linear',
+    },
+    y: {
+      bounds: ['', ''],
+      label: '',
+      prefix: '',
+      suffix: '',
+      base: '10',
+      scale: 'linear',
+    },
+  },
+  geom: 'line',
+  shape: 'chronograf-v2',
+}
