+++ src/notebooks/pipes/Visualization/types/Mosaic/index.tsx
@@ -1,194 +1,5 @@
-import React from 'react'
-import {NINETEEN_EIGHTY_FOUR} from '@influxdata/giraffe'
-import {
-  DEFAULT_FILLVALUES,
-  AGG_WINDOW_AUTO,
-} from 'src/timeMachine/constants/queryBuilder'
-import {MosaicViewProperties} from 'src/types'
-
-const icon = (
-  <div className="vis-graphic" data-testid="vis-graphic--mosaic">
-    <svg
-      width="100%"
-      height="100%"
-      version="1.1"
-      id="Bar"
-      x="0px"
-      y="0px"
-      viewBox="0 0 150 150"
-      preserveAspectRatio="none meet"
-    >
-      <rect
-        x="96"
-        y="58"
-        className="vis-graphic--line vis-graphic--line-b"
-        width="44"
-        height="34"
-      />
-      <rect
-        x="57"
-        y="95"
-        className="vis-graphic--line vis-graphic--line-b"
-        width="50"
-        height="33"
-      />
-      <rect
-        x="78"
-        y="22"
-        className="vis-graphic--line vis-graphic--line-b"
-        width="26"
-        height="33"
-      />
-      <rect
-        x="96"
-        y="58"
-        className="vis-graphic--fill vis-graphic--fill-b"
-        width="44"
-        height="34"
-      />
-      <rect
-        x="57"
-        y="95"
-        className="vis-graphic--fill vis-graphic--fill-b"
-        width="50"
-        height="33"
-      />
-      <rect
-        x="78"
-        y="22"
-        className="vis-graphic--fill vis-graphic--fill-b"
-        width="26"
-        height="33"
-      />
-      <rect
-        x="67"
-        y="22"
-        className="vis-graphic--line vis-graphic--line-c"
-        width="10"
-        height="33"
-      />
-      <rect
-        x="105"
-        y="22"
-        className="vis-graphic--line vis-graphic--line-c"
-        width="14"
-        height="33"
-      />
-      <rect
-        x="10"
-        y="95"
-        className="vis-graphic--line vis-graphic--line-c"
-        width="23"
-        height="33"
-      />
-      <rect
-        x="24"
-        y="58"
-        className="vis-graphic--line vis-graphic--line-c"
-        width="71"
-        height="34"
-      />
-      <rect
-        x="67"
-        y="22"
-        className="vis-graphic--fill vis-graphic--fill-c"
-        width="10"
-        height="33"
-      />
-      <rect
-        x="105"
-        y="22"
-        className="vis-graphic--fill vis-graphic--fill-c"
-        width="14"
-        height="33"
-      />
-      <rect
-        x="10"
-        y="95"
-        className="vis-graphic--fill vis-graphic--fill-c"
-        width="23"
-        height="33"
-      />
-      <rect
-        x="24"
-        y="58"
-        className="vis-graphic--fill vis-graphic--fill-c"
-        width="71"
-        height="34"
-      />
-      <rect
-        x="120"
-        y="22"
-        className="vis-graphic--line vis-graphic--line-a"
-        width="20"
-        height="33"
-      />
-      <rect
-        x="10"
-        y="22"
-        className="vis-graphic--line vis-graphic--line-a"
-        width="56"
-        height="33"
-      />
-      <rect
-        x="34"
-        y="95"
-        className="vis-graphic--line vis-graphic--line-a"
-        width="22"
-        height="33"
-      />
-      <rect
-        x="108"
-        y="95"
-        className="vis-graphic--line vis-graphic--line-a"
-        width="32"
-        height="33"
-      />
-      <rect
-        x="10"
-        y="58"
-        className="vis-graphic--line vis-graphic--line-a"
-        width="13"
-        height="34"
-      />
-      <rect
-        x="120"
-        y="22"
-        className="vis-graphic--fill vis-graphic--fill-a"
-        width="20"
-        height="33"
-      />
-      <rect
-        x="10"
-        y="22"
-        className="vis-graphic--fill vis-graphic--fill-a"
-        width="56"
-        height="33"
-      />
-      <rect
-        x="34"
-        y="95"
-        className="vis-graphic--fill vis-graphic--fill-a"
-        width="22"
-        height="33"
-      />
-      <rect
-        x="108"
-        y="95"
-        className="vis-graphic--fill vis-graphic--fill-a"
-        width="32"
-        height="33"
-      />
-      <rect
-        x="10"
-        y="58"
-        className="vis-graphic--fill vis-graphic--fill-a"
-        width="13"
-        height="34"
-      />
-    </svg>
-  </div>
-)
+import icon from './icon'
+import properties from './properties'
 
 export default register => {
   register({
@@ -196,47 +7,6 @@ export default register => {
     name: 'Mosaic',
     graphic: icon,
     featureFlag: 'mosaicGraphType',
-    initial: {
-      type: 'mosaic',
-      shape: 'chronograf-v2',
-
-      queries: [
-        {
-          name: '',
-          text: '',
-          editMode: 'builder',
-          builderConfig: {
-            buckets: [],
-            tags: [
-              {
-                key: '_measurement',
-                values: [],
-                aggregateFunctionType: 'filter',
-              },
-            ],
-            functions: [{name: 'mean'}],
-            aggregateWindow: {
-              period: AGG_WINDOW_AUTO,
-              fillValues: DEFAULT_FILLVALUES,
-            },
-          },
-        },
-      ],
-
-      colors: NINETEEN_EIGHTY_FOUR,
-      note: '',
-      showNoteWhenEmpty: false,
-      fillColumns: null,
-      xColumn: null,
-      xDomain: null,
-      ySeriesColumns: null,
-      yDomain: null,
-      xAxisLabel: '',
-      yAxisLabel: '',
-      xPrefix: '',
-      xSuffix: '',
-      yPrefix: '',
-      ySuffix: '',
-    } as MosaicViewProperties,
+    initial: properties,
   })
 }
