+++ src/notebooks/pipes/Visualization/types/SingleStatWithLine/properties.ts
@@ -1,0 +1,69 @@
+import {DEFAULT_THRESHOLDS_LIST_COLORS} from 'src/shared/constants/thresholds'
+import {
+  DEFAULT_FILLVALUES,
+  AGG_WINDOW_AUTO,
+} from 'src/timeMachine/constants/queryBuilder'
+
+import {Axis, Base, Color, LinePlusSingleStatProperties} from 'src/types'
+
+export default {
+  type: 'line-plus-single-stat',
+  shape: 'chronograf-v2',
+  xColumn: null,
+  yColumn: null,
+  position: 'overlaid',
+
+  queries: [
+    {
+      name: '',
+      text: '',
+      editMode: 'builder',
+      builderConfig: {
+        buckets: [],
+        tags: [
+          {
+            key: '_measurement',
+            values: [],
+            aggregateFunctionType: 'filter',
+          },
+        ],
+        functions: [{name: 'mean'}],
+        aggregateWindow: {
+          period: AGG_WINDOW_AUTO,
+          fillValues: DEFAULT_FILLVALUES,
+        },
+      },
+    },
+  ],
+
+  legend: {},
+  note: '',
+  showNoteWhenEmpty: false,
+  axes: {
+    x: {
+      bounds: ['', ''],
+      label: '',
+      prefix: '',
+      suffix: '',
+      base: '10',
+      scale: 'linear',
+    } as Axis,
+    y: {
+      bounds: ['', ''],
+      label: '',
+      prefix: '',
+      suffix: '',
+      base: '10' as Base,
+      scale: 'linear',
+    } as Axis,
+  },
+  colors: DEFAULT_THRESHOLDS_LIST_COLORS as Color[],
+  prefix: '',
+  tickPrefix: '',
+  suffix: '',
+  tickSuffix: '',
+  decimalPlaces: {
+    isEnforced: true,
+    digits: 2,
+  },
+} as LinePlusSingleStatProperties
