+++ src/notebooks/pipes/Visualization/types/Table/icon.tsx
@@ -1,0 +1,80 @@
+import React from 'react'
+
+const icon = (
+  <div className="vis-graphic" data-testid="vis-graphic--table">
+    <svg
+      id="Table"
+      x="0px"
+      y="0px"
+      width="100%"
+      height="100%"
+      viewBox="0 0 150 150"
+    >
+      <path
+        className="vis-graphic--fill vis-graphic--fill-c"
+        d="M55.5,115H19.7c-1.7,0-3.1-1.4-3.1-3.1V61.7h38.9V115z"
+      />
+      <path
+        className="vis-graphic--fill vis-graphic--fill-b"
+        d="M133.4,61.7H55.5V35h74.8c1.7,0,3.1,1.4,3.1,3.1V61.7z"
+      />
+      <path
+        className="vis-graphic--fill vis-graphic--fill-a"
+        d="M55.5,61.7H16.6V38.1c0-1.7,1.4-3.1,3.1-3.1h35.9V61.7z"
+      />
+      <path
+        className="vis-graphic--line vis-graphic--line-c"
+        d="M16.6,88.3v23.6c0,1.7,1.4,3.1,3.1,3.1h35.9V88.3H16.6z"
+      />
+      <rect
+        className="vis-graphic--line vis-graphic--line-c"
+        x="16.6"
+        y="61.7"
+        width="38.9"
+        height="26.7"
+      />
+      <path
+        className="vis-graphic--line vis-graphic--line-b"
+        d="M94.5,35v26.7h38.9V38.1c0-1.7-1.4-3.1-3.1-3.1H94.5z"
+      />
+      <rect
+        className="vis-graphic--line vis-graphic--line-b"
+        x="55.5"
+        y="35"
+        width="38.9"
+        height="26.7"
+      />
+      <path
+        className="vis-graphic--line vis-graphic--line-d"
+        d="M94.5,115h35.9c1.7,0,3.1-1.4,3.1-3.1V88.3H94.5V115z"
+      />
+      <rect
+        className="vis-graphic--line vis-graphic--line-d"
+        x="55.5"
+        y="88.3"
+        width="38.9"
+        height="26.7"
+      />
+      <rect
+        className="vis-graphic--line vis-graphic--line-d"
+        x="94.5"
+        y="61.7"
+        width="38.9"
+        height="26.7"
+      />
+      <rect
+        className="vis-graphic--line vis-graphic--line-d"
+        x="55.5"
+        y="61.7"
+        width="38.9"
+        height="26.7"
+      />
+      <path
+        className="vis-graphic--line vis-graphic--line-a"
+        d="M55.5,35H19.7c-1.7,0-3.1,1.4-3.1,3.1v23.6h38.9V35z"
+      />
+    </svg>
+  </div>
+)
+
+export default icon
